mostly, get rid of the 'rand' indexing, which was nonsense.
it totally broke everything; my tables were never getting any real
data in them.  so, now, everybody gets a chunk of the threads in
the request, and the process that block sequentially.

i upped the thread count on the kernels.  this now means we get an
occupancy of 1.0, so we're actually using all the hardware. woo.
this necessitated fiddling with the size of the cache: it seems it
has issues unless the local cache is the same size as the number
of threads.

i wrote some code to take a look at the values in the HT after a
successful kernel completion.  next step is to iterate until the
requests list clears, like a real progressive renderer would.

added a 'verbose' option to avoid spam / IO in normal operation
without always commenting it out.  nice for switching between
benchmarking && debugging.