fixed up the HT script.
First of all, it's actually slow now, so I do a much sparser sampling.
Secondly, I check for kernel launch failures and don't record data in
that case; this will happen with small HTs, which is why we now start
at 25 instead of 1; values lower than 25 fail *often*; a 1-element HT
seems to fail always.
'fail' is a misnomer here; we hit a timeout.  Maybe there's a way
to increase the timeout?