Iterate until convergence; ability to load BA files.
Iteration seems to be partially broken; mostly works for the naive
case, and fails often in the other case.

Also, our hash tables don't seem to be filling consistently; this
may be an issue with the input dataset, however.

Secondly, integrated code from Alex to read '.ba' files, which
describe the set of brick requests in a run.  These are actually
pulled from live IV3D runs, so produce much more realistic data.