Fix the bug iterating/deleting entries from the HT.
The issue actually turned out to be the incorrect marking of
remove_entry as a PURE function.  PURE is only for functions which
produce a value; any function which modifies its argument
can/should not be pure.

Oops.

Also managed to break the non-naive based kernel, somehow.  It
seems to never copy any values into the global HT mem...